server/requirements.txt
fastapi
uvicorn
pandas
numpy
fairlearn
google-generativeai
python-dotenv
pydantic
scikit-learn
